[RUM-16084] Merge remote config classes, refactor URL construction, move lifecycle wiring into DatadogCore
